fix: allow finalizer to cleanup to continue on error (fixes #173)
currently finalizer cleanup will never continue when this._getSdkKey errors. In this case, a keyref pointing to a key/secret that doesnt exist will cause finalizer cleanup to fail forever. I considered narrowing the scope of error that we allow to be skipped, but for now am just leaving it wide open until more time is availbe to decide on a better subset of error codes that we should allow skipping on. I believe the only side effects of this, is after a while, if there are a lot of keys that dont get cleaned up, we end up leaving open a lot of open conections to LD. In most cases this shouldnt be an issue since it is rare to use a bunch of different sdk keys, and even rarer to try to clean them all up and have issues before the pod gets restarted and all connections are closed anyways.