fix: allow finalizer to continue cleanup when error (fixes #173) (#174)
* fix: allow finalizer to cleanup to continue on error (fixes #173)

currently finalizer cleanup will never continue when this._getSdkKey errors. In this case, a keyref pointing to a key/secret that doesnt exist will cause finalizer cleanup to fail forever.

* fix: improve sdkkey lookup when it fails to get it from the cluster

this is a backup solution for when getting the sdkkey directly fails by using the instanceUid to figure out which sdkkey it is associated with. ASsuming there arent ever that many sdkkeys, this shouldnt become to computationally expensive.

* Update FeatureFlagSetLDController.js

* Update src/FeatureFlagSetLDController.js

Co-authored-by: Adam King <rak@linux.vnet.ibm.com>

* Update src/FeatureFlagSetLDController.js

Co-authored-by: Adam King <rak@linux.vnet.ibm.com>

* improve code readability

Co-authored-by: Adam King <rak@linux.vnet.ibm.com>